backend/requirements.txt
fastapi
uvicorn
motor
beanie
bcrypt
python-jose[cryptography]
python-dotenv
scikit-learn
